+++ packages/components/src/tests/commonTests/browser/interfaces.ts
@@ -1,0 +1,5 @@
+import { mount } from "@arcgis/lumina-compiler/testing";
+
+export interface TestSetup {
+  (): ReturnType<typeof mount>;
+}